Unify logging: replace logging_linux.h + macOS stubs with shared logging.h
logging_linux.h had nothing Linux-specific — pure fprintf/cstdio. Collapse
it and the SDL switcher in logging.h into a single cross-platform header,
removing the inline macOS stub block from main.cpp.
